BackEnd/requirements.txt
flask
flask_cors
bcrypt
python-dotenv
google-generativeai